# docs/openapi.yml
openapi: 3.1.0
info:
    title: Java API MVC With Postgres - Extension Criteria
    description: Extension API for Java API with Postgres Exercise
    version: 0.2.3
    contact:
        name: MVC with Postgres Exercise

servers:
    - url: http://localhost:4000
tags:
    - name: employees
    - name: salaries
    - name: departments
paths:
    /employees:
        post:
            tags:
                - employees
            summary: Create a new employee
            description: Add a new employee to our list of available employees
            operationId: createEmployee
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateEmployee'
            responses:
                '201':
                    description: Successfully created a new employee
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Employee'
                '400':
                    description: "Could not create employee, please check all required fields are correct."
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
        get:
            tags:
                - employees
            summary: Get all employees
            description: Get all of the employees from our list
            operationId: getEmployees
            responses:
                '200':
                    description: Successfully returned a list of all employees
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllEmployees'

    /employees/{id}:
        get:
            tags:
                - employees
            summary: Get an employee
            description: Get an employee by ID
            operationId: getSpecificEmployee
            parameters:
                - name: id
                  in: path
                  description: The ID of the employee to get
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Successfully returned the specified employee
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Employee'
                '404':
                    description: "No employees with that id were found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        put:
            tags:
                - employees
            summary: Update an employee
            description: Update an existing employee by ID
            operationId: updateEmployee
            parameters:
                - name: id
                  in: path
                  description: The ID of the employee to update
                  required: true
                  schema:
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateEmployee'
            responses:
                '201':
                    description: Successfully updated the specified employee
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Employee'
                '400':
                    description: "Could not update the employee, please check all required fields are correct."
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
                '404':
                    description: "No employee with that id was found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        delete:
            tags:
                - employees
            summary: Delete an employee
            description: Remove an employee from our list of employees
            operationId: deleteEmployee
            parameters:
                - name: id
                  in: path
                  description: The ID of the employee to delete
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Successfully deleted the specified employee
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Employee'
                '404':
                    description: "No employee with that id was found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
    /salaries:
        post:
            tags:
                - salaries
            summary: Create a new salary grade
            description: Add a new salary grade to our list of available salary grades
            operationId: createSalary
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateSalary'
            responses:
                '201':
                    description: Successfully created a new salary grade
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Salary'
                '400':
                    description: "Could not create a new salary grade, please check all required fields are correct."
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
        get:
            tags:
                - salaries
            summary: Get all salary grades
            description: Get all of the salary grades from our list
            operationId: getSalaries
            responses:
                '200':
                    description: Successfully returned all of the salary grades
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllSalaries'

    /salaries/{grade}:
        get:
            tags:
                - salaries
            summary: Get a salary grade
            description: Get a salary grade
            operationId: getSpecificSalary
            parameters:
                - name: grade
                  in: path
                  description: The salary grade to get
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Returned the specific salary
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Salary'
                '404':
                    description: "No salary grades matching that id were found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        put:
            tags:
                - salaries
            summary: Update a salary grade
            description: Update an existing salary grade
            operationId: updateSalary
            parameters:
                - name: grade
                  in: path
                  description: The salary grade to update
                  required: true
                  schema:
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateSalary'
            responses:
                '201':
                    description: Successfully updated the specified salary grade
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Salary'
                '400':
                    description: "Could not update the salary grade, please check all required fields are correct."
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
                '404':
                    description: "No salary grades matching that id were found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        delete:
            tags:
                - salaries
            summary: Delete a salary grade
            description: Remove a salary grade from our list of salary grades
            operationId: deleteSalary
            parameters:
                - name: grade
                  in: path
                  description: The salary grade to delete
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Successfully deleted the specified salary grade
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Salary'
                '404':
                    description: No salary grade was found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
    /departments:
        post:
            tags:
                - departments
            summary: Create a new department
            description: Add a new department to our list of available departments
            operationId: createDepartment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateDepartment'
            responses:
                '201':
                    description: Successfully created the specified department
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Department'
                '400':
                    description: Could not create the new department, please check all required fields are correct.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
        get:
            tags:
                - departments
            summary: Get all departments
            description: Get all of the departments from our list
            operationId: getDepartments
            responses:
                '200':
                    description: Successfully returned a list of all departments
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AllDepartments'

    /departments/{id}:
        get:
            tags:
                - departments
            summary: Get a department
            description: Get a department by ID
            operationId: getSpecificDepartment
            parameters:
                - name: id
                  in: path
                  description: The ID of the department to get
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Successfully returned the specified Department
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Department'
                '404':
                    description: No departments matching that id were found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        put:
            tags:
                - departments
            summary: Update a department
            description: Update an existing department by ID
            operationId: updateDepartment
            parameters:
                - name: id
                  in: path
                  description: The ID of the department to update
                  required: true
                  schema:
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateDepartment'
            responses:
                '201':
                    description: Successfully updated the specified department
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Department'
                '400':
                    description: Could not update the department, please check all required fields are correct.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error400'
                '404':
                    description: No departments matching that id were found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'
        delete:
            tags:
                - departments
            summary: Delete a department
            description: Remove a department from our list of departments
            operationId: deleteDepartment
            parameters:
                - name: id
                  in: path
                  description: The ID of the department to delete
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Successfully deleted the specified department
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Department'
                '404':
                    description: "No departments matching that id were found"
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error404'

components:
    schemas:
        CreateEmployee:
            type: object
            required: [name, jobName, salaryGrade, departmentId]
            properties:
                name:
                    type: string
                    example: "Dave Ames"
                jobName:
                    type: string
                    example: "Java Developer"
                salaryGrade:
                    type: integer
                    example: 8
                department:
                    type: integer
                    example: 2
        Employee:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                name:
                    type: string
                    example: "Dave Ames"
                jobName:
                    type: string
                    example: "Java Developer"
                salaryGrade:
                    type: integer
                    example: 8
                department:
                    type: integer
                    example: 2
        AllEmployees:
            type: array
            items:
                $ref: '#/components/schemas/Employee'
        CreateSalary:
            type: object
            required: [grade, minSalary, maxSalary]
            properties:
                grade:
                    type: integer
                    example: 10
                minSalary:
                    type: integer
                    example: 25000
                maxSalary:
                    type: integer
                    example: 30000
        Salary:
            type: object
            properties:
                grade:
                    type: integer
                    example: 10
                minSalary:
                    type: integer
                    example: 25000
                maxSalary:
                    type: integer
                    example: 30000
        AllSalaries:
            type: array
            items:
                $ref: '#/components/schemas/Salary'

        CreateDepartment:
            type: object
            required: [name, location]
            properties:
                name:
                    type: string
                    example: "Software Development"
                location:
                    type: string
                    example: "Manchester, UK"
        Department:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                name:
                    type: string
                    example: "Software Development"
                location:
                    type: string
                    example: "Manchester, UK"
        AllDepartments:
            type: array
            items:
                $ref: '#/components/schemas/Department'
        Error400:
            type: object
            properties:
                message:
                    type: string
                    example: "bad request"
        Error404:
            type: object
            properties:
                message:
                    type: string
                    example: "not found"